Default data table camel case converter
