+++ java/src/test/java/io/cucumber/java/defaultstransformer/DataTableSteps.java
@@ -20,7 +20,7 @@ public class DataTableSteps {
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     @DefaultParameterTransformer
-    @DefaultDataTableEntryTransformer
+    @DefaultDataTableEntryTransformer(titleCaseHeaders = true)
     @DefaultDataTableCellTransformer
     public Object defaultTransformer(Object fromValue, Type toValueType) {
         return objectMapper.convertValue(fromValue, objectMapper.constructType(toValueType));
@@ -28,6 +28,11 @@ public class DataTableSteps {
 
     @Given("a list of authors in a table")
     public void aListOfAuthorsInATable(List<Author> authors) {
+        assertTrue(authors.contains(expectedAuthor));
+    }
+
+    @Given("a table with capital case headers")
+    public void aTableWithCapitalCaseHeaders(List<Author> authors) {
         assertTrue(authors.contains(expectedAuthor));
     }
